# [exporter/splunkhec] Introduce pools for buffers states