--- a/.chloggen/splunkhec-improve-performance.yaml
+++ b/.chloggen/splunkhec-improve-performance.yaml
@@ -1,0 +1,11 @@
+# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
+change_type: enhancement
+
+# The name of the component, or a single word describing the area of concern, (e.g. filelogreceiver)
+component: exporter/splunkhec
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: Improve performance and reduce memory consumption.
+
+# One or more tracking issues related to the change
+issues: [21765, 21791]