--- a/exporter/splunkhecexporter/client_test.go
+++ b/exporter/splunkhecexporter/client_test.go
@@ -44,7 +44,6 @@ import (
 	"go.opentelemetry.io/collector/pdata/ptrace"
 	conventions "go.opentelemetry.io/collector/semconv/v1.6.1"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zaptest"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/splunk"
 )
@@ -1033,11 +1032,8 @@ func Test_PushMetricsData_Histogram_NaN_Sum(t *testing.T) {
 	dp := histogram.SetEmptyHistogram().DataPoints().AppendEmpty()
 	dp.SetSum(math.NaN())
 
-	c := client{
-		config:    NewFactory().CreateDefaultConfig().(*Config),
-		logger:    zap.NewNop(),
-		hecWorker: &mockHecWorker{},
-	}
+	c := newMetricsClient(exportertest.NewNopCreateSettings(), NewFactory().CreateDefaultConfig().(*Config))
+	c.hecWorker = &mockHecWorker{}
 
 	permanentErrors := c.pushMetricsDataInBatches(context.Background(), metrics, map[string]string{})
 	assert.NoError(t, permanentErrors)
@@ -1052,11 +1048,8 @@ func Test_PushMetricsData_Summary_NaN_Sum(t *testing.T) {
 	dp := summary.SetEmptySummary().DataPoints().AppendEmpty()
 	dp.SetSum(math.NaN())
 
-	c := client{
-		config:    NewFactory().CreateDefaultConfig().(*Config),
-		logger:    zap.NewNop(),
-		hecWorker: &mockHecWorker{},
-	}
+	c := newMetricsClient(exportertest.NewNopCreateSettings(), NewFactory().CreateDefaultConfig().(*Config))
+	c.hecWorker = &mockHecWorker{}
 
 	permanentErrors := c.pushMetricsDataInBatches(context.Background(), metrics, map[string]string{})
 	assert.NoError(t, permanentErrors)
@@ -1213,10 +1206,7 @@ func Test_pushLogData_nil_Logs(t *testing.T) {
 		},
 	}
 
-	c := client{
-		config: NewFactory().CreateDefaultConfig().(*Config),
-		logger: zaptest.NewLogger(t),
-	}
+	c := newLogsClient(exportertest.NewNopCreateSettings(), NewFactory().CreateDefaultConfig().(*Config))
 
 	for _, test := range tests {
 		for _, disabled := range []bool{true, false} {
@@ -1231,10 +1221,7 @@ func Test_pushLogData_nil_Logs(t *testing.T) {
 }
 
 func Test_pushLogData_InvalidLog(t *testing.T) {
-	c := client{
-		config: NewFactory().CreateDefaultConfig().(*Config),
-		logger: zaptest.NewLogger(t),
-	}
+	c := newLogsClient(exportertest.NewNopCreateSettings(), NewFactory().CreateDefaultConfig().(*Config))
 
 	logs := plog.NewLogs()
 	log := logs.ResourceLogs().AppendEmpty().ScopeLogs().AppendEmpty().LogRecords().AppendEmpty()
@@ -1247,11 +1234,8 @@ func Test_pushLogData_InvalidLog(t *testing.T) {
 }
 
 func Test_pushLogData_PostError(t *testing.T) {
-	c := client{
-		config:    NewFactory().CreateDefaultConfig().(*Config),
-		logger:    zaptest.NewLogger(t),
-		hecWorker: &defaultHecWorker{url: &url.URL{Host: "in va lid"}},
-	}
+	c := newLogsClient(exportertest.NewNopCreateSettings(), NewFactory().CreateDefaultConfig().(*Config))
+	c.hecWorker = &defaultHecWorker{url: &url.URL{Host: "in va lid"}}
 
 	// 2000 log records -> ~371888 bytes when JSON encoded.
 	logs := createLogData(1, 1, 2000)
@@ -1289,10 +1273,7 @@ func Test_pushLogData_PostError(t *testing.T) {
 func Test_pushLogData_ShouldAddResponseTo400Error(t *testing.T) {
 	config := NewFactory().CreateDefaultConfig().(*Config)
 	url := &url.URL{Scheme: "http", Host: "splunk"}
-	splunkClient := client{
-		config: config,
-		logger: zaptest.NewLogger(t),
-	}
+	splunkClient := newLogsClient(exportertest.NewNopCreateSettings(), NewFactory().CreateDefaultConfig().(*Config))
 	logs := createLogData(1, 1, 1)
 
 	responseBody := `some error occurred`
@@ -1320,17 +1301,15 @@ func Test_pushLogData_ShouldAddResponseTo400Error(t *testing.T) {
 
 func Test_pushLogData_ShouldReturnUnsentLogsOnly(t *testing.T) {
 	config := NewFactory().CreateDefaultConfig().(*Config)
+
+	// Each record is about 200 bytes, so the 250-byte buffer will fit only one at a time
+	config.MaxContentLengthLogs, config.DisableCompression = 250, true
+
 	url := &url.URL{Scheme: "http", Host: "splunk"}
-	c := client{
-		config: config,
-		logger: zaptest.NewLogger(t),
-	}
+	c := newLogsClient(exportertest.NewNopCreateSettings(), config)
 
 	// Just two records
 	logs := createLogData(2, 1, 1)
-
-	// Each record is about 200 bytes, so the 250-byte buffer will fit only one at a time
-	c.config.MaxContentLengthLogs, c.config.DisableCompression = 250, true
 
 	// The first record is to be sent successfully, the second one should not
 	httpClient, _ := newTestClientWithPresetResponses([]int{200, 400}, []string{"OK", "NOK"})
@@ -1349,20 +1328,18 @@ func Test_pushLogData_ShouldReturnUnsentLogsOnly(t *testing.T) {
 
 func Test_pushLogData_ShouldAddHeadersForProfilingData(t *testing.T) {
 	config := NewFactory().CreateDefaultConfig().(*Config)
-	url := &url.URL{Scheme: "http", Host: "splunk"}
-	c := client{
-		config: config,
-		logger: zaptest.NewLogger(t),
-	}
+
+	// A 300-byte buffer only fits one record (around 200 bytes), so each record will be sent separately
+	config.MaxContentLengthLogs, config.DisableCompression = 300, true
+
+	c := newLogsClient(exportertest.NewNopCreateSettings(), config)
 
 	logs := createLogDataWithCustomLibraries(1, []string{"otel.logs", "otel.profiling"}, []int{10, 20})
 	var headers *[]http.Header
 
 	httpClient, headers := newTestClient(200, "OK")
+	url := &url.URL{Scheme: "http", Host: "splunk"}
 	c.hecWorker = &defaultHecWorker{url, httpClient, buildHTTPHeaders(config, component.NewDefaultBuildInfo())}
-
-	// A 300-byte buffer only fits one record (around 200 bytes), so each record will be sent separately
-	c.config.MaxContentLengthLogs, c.config.DisableCompression = 300, true
 
 	err := c.pushLogData(context.Background(), logs)
 	require.NoError(t, err)
@@ -1412,22 +1389,34 @@ func Benchmark_pushLogData_10_10_10_8K(b *testing.B) {
 func Benchmark_pushLogData_10_10_10_1M(b *testing.B) {
 	benchPushLogData(b, 10, 10, 10, 1024*1024)
 }
+
+func Benchmark_pushLogData_5_0_20_2M(b *testing.B) {
+	benchPushLogData(b, 5, 0, 20, 2*1024*1024)
+}
+
+func Benchmark_pushLogData_10_0_100_2M(b *testing.B) {
+	benchPushLogData(b, 10, 0, 100, 2*1024*1024)
+}
+
+func Benchmark_pushLogData_50_0_200_2M(b *testing.B) {
+	benchPushLogData(b, 50, 0, 200, 2*1024*1024)
+}
+
+func Benchmark_pushLogData_50_0_200_5M(b *testing.B) {
+	benchPushLogData(b, 50, 0, 200, 5*1024*1024)
+}
+
 func Benchmark_pushLogData_10_1_1_1024(b *testing.B) {
 	benchPushLogData(b, 10, 1, 1, 1024)
 }
 
 func benchPushLogData(b *testing.B, numResources int, numProfiling int, numNonProfiling int, bufSize uint) {
 	config := NewFactory().CreateDefaultConfig().(*Config)
-	url := &url.URL{Scheme: "http", Host: "splunk"}
-	c := client{
-		config: config,
-		logger: zaptest.NewLogger(b),
-	}
+	config.MaxContentLengthLogs = bufSize
+	config.DisableCompression = true
+	c := newLogsClient(exportertest.NewNopCreateSettings(), config)
+	c.hecWorker = &mockHecWorker{}
 
-	httpClient, _ := newTestClient(200, "OK")
-	c.hecWorker = &defaultHecWorker{url, httpClient, buildHTTPHeaders(config, component.NewDefaultBuildInfo())}
-
-	c.config.MaxContentLengthLogs = bufSize
 	logs := createLogDataWithCustomLibraries(numResources, []string{"otel.logs", "otel.profiling"}, []int{numNonProfiling, numProfiling})
 
 	b.ReportAllocs()
@@ -1441,17 +1430,15 @@ func benchPushLogData(b *testing.B, numResources int, numProfiling int, numNonPr
 
 func Test_pushLogData_Small_MaxContentLength(t *testing.T) {
 	config := NewFactory().CreateDefaultConfig().(*Config)
-	c := client{
-		config:    config,
-		logger:    zaptest.NewLogger(t),
-		hecWorker: &defaultHecWorker{&url.URL{Scheme: "http", Host: "splunk"}, http.DefaultClient, buildHTTPHeaders(config, component.NewDefaultBuildInfo())},
-	}
-	c.config.MaxContentLengthLogs = 1
+	config.MaxContentLengthLogs = 1
 
 	logs := createLogData(1, 1, 2000)
 
 	for _, disable := range []bool{true, false} {
-		c.config.DisableCompression = disable
+		config.DisableCompression = disable
+
+		c := newLogsClient(exportertest.NewNopCreateSettings(), config)
+		c.hecWorker = &defaultHecWorker{&url.URL{Scheme: "http", Host: "splunk"}, http.DefaultClient, buildHTTPHeaders(config, component.NewDefaultBuildInfo())}
 
 		err := c.pushLogData(context.Background(), logs)
 		require.Error(t, err)
@@ -1590,9 +1577,10 @@ func TestPushLogRecordsBufferCounters(t *testing.T) {
 	cfg := NewFactory().CreateDefaultConfig().(*Config)
 	cfg.ExportRaw = true
 	c := client{
-		config:    cfg,
-		logger:    zap.NewNop(),
-		hecWorker: &mockHecWorker{},
+		config:          cfg,
+		logger:          zap.NewNop(),
+		hecWorker:       &mockHecWorker{},
+		bufferStatePool: newBufferStatePool(6, false, 10),
 	}
 
 	logs := plog.NewResourceLogsSlice()
@@ -1600,7 +1588,8 @@ func TestPushLogRecordsBufferCounters(t *testing.T) {
 	logRecords.AppendEmpty().Body().SetStr("12345")    // the first log record should be accepted and sent
 	logRecords.AppendEmpty().Body().SetStr("12345678") // the second log record should be rejected as it's too big
 	logRecords.AppendEmpty().Body().SetStr("12345")    // the third log record should be just accepted
-	bs := makeBlankBufferState(6, false, 10)
+	bs := c.bufferStatePool.get()
+	defer c.bufferStatePool.put(bs)
 
 	permErrs, err := c.pushLogRecords(context.Background(), logs, bs, nil)
 	assert.NoError(t, err)
@@ -1627,15 +1616,16 @@ func validateCompressedContains(t *testing.T, expected []string, got []byte) {
 func BenchmarkPushLogRecords(b *testing.B) {
 	logs := createLogData(1, 1, 1)
 	c := client{
-		config:    NewFactory().CreateDefaultConfig().(*Config),
-		logger:    zap.NewNop(),
-		hecWorker: &mockHecWorker{},
+		config:          NewFactory().CreateDefaultConfig().(*Config),
+		logger:          zap.NewNop(),
+		hecWorker:       &mockHecWorker{},
+		bufferStatePool: newBufferStatePool(4096, true, 4096),
 	}
 
 	b.ReportAllocs()
 	b.ResetTimer()
 
-	state := makeBlankBufferState(4096, true, 4096)
+	state := c.bufferStatePool.get()
 	for n := 0; n < b.N; n++ {
 		permanentErrs, sendingErr := c.pushLogRecords(context.Background(), logs.ResourceLogs(), state, map[string]string{})
 		assert.NoError(b, sendingErr)