To improve performance and reduce memory consumption. 

Mostly applicable to configurations with disabled encryption.

Before:

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
Benchmark_pushLogData_100_10_10_1024
Benchmark_pushLogData_100_10_10_1024-10     	     544	   2265471 ns/op	 1834813 B/op	   26147 allocs/op
Benchmark_pushLogData_10_100_100_1024
Benchmark_pushLogData_10_100_100_1024-10    	     538	   2251496 ns/op	 1836345 B/op	   26327 allocs/op
Benchmark_pushLogData_10_0_100_1024
Benchmark_pushLogData_10_0_100_1024-10      	    1158	   1044071 ns/op	  923662 B/op	   13168 allocs/op
Benchmark_pushLogData_10_100_0_1024
Benchmark_pushLogData_10_100_0_1024-10      	    1168	   1051447 ns/op	  923645 B/op	   13168 allocs/op
Benchmark_pushLogData_10_10_10_256
Benchmark_pushLogData_10_10_10_256-10       	    6034	    203278 ns/op	  192722 B/op	    2789 allocs/op
Benchmark_pushLogData_10_10_10_1024
Benchmark_pushLogData_10_10_10_1024-10      	    6188	    196455 ns/op	  191089 B/op	    2623 allocs/op
Benchmark_pushLogData_10_10_10_8K
Benchmark_pushLogData_10_10_10_8K-10        	    6066	    194614 ns/op	  204441 B/op	    2593 allocs/op
Benchmark_pushLogData_10_10_10_1M
Benchmark_pushLogData_10_10_10_1M-10        	    2432	    449211 ns/op	 2294672 B/op	    2593 allocs/op
Benchmark_pushLogData_5_0_20_2M
Benchmark_pushLogData_5_0_20_2M-10          	    3075	    346789 ns/op	 4303053 B/op	    1308 allocs/op
Benchmark_pushLogData_10_0_100_2M
Benchmark_pushLogData_10_0_100_2M-10        	     825	   1382169 ns/op	 5122079 B/op	   13005 allocs/op
Benchmark_pushLogData_50_0_200_2M
Benchmark_pushLogData_50_0_200_2M-10        	     100	  10067283 ns/op	13329647 B/op	  129966 allocs/op
Benchmark_pushLogData_50_0_200_5M
Benchmark_pushLogData_50_0_200_5M-10        	     100	  10150482 ns/op	19620303 B/op	  129964 allocs/op
Benchmark_pushLogData_10_1_1_1024
Benchmark_pushLogData_10_1_1_1024-10        	   61316	     20100 ns/op	   29232 B/op	     253 allocs/op
BenchmarkPushLogRecords
BenchmarkPushLogRecords-10                  	  213103	      5393 ns/op	   31409 B/op	      13 allocs/op
PASS
```


After:
```
Benchmark_pushLogData_100_10_10_1024
Benchmark_pushLogData_100_10_10_1024-10     	     540	   2153153 ns/op	 1819994 B/op	   25813 allocs/op
Benchmark_pushLogData_10_100_100_1024
Benchmark_pushLogData_10_100_100_1024-10    	     541	   2295616 ns/op	 1821203 B/op	   25992 allocs/op
Benchmark_pushLogData_10_0_100_1024
Benchmark_pushLogData_10_0_100_1024-10      	    1162	   1035513 ns/op	  910593 B/op	   12998 allocs/op
Benchmark_pushLogData_10_100_0_1024
Benchmark_pushLogData_10_100_0_1024-10      	    1164	   1030327 ns/op	  910577 B/op	   12998 allocs/op
Benchmark_pushLogData_10_10_10_256
Benchmark_pushLogData_10_10_10_256-10       	    6300	    191156 ns/op	  179660 B/op	    2585 allocs/op
Benchmark_pushLogData_10_10_10_1024
Benchmark_pushLogData_10_10_10_1024-10      	    6360	    189267 ns/op	  179666 B/op	    2585 allocs/op
Benchmark_pushLogData_10_10_10_8K
Benchmark_pushLogData_10_10_10_8K-10        	    6318	    188939 ns/op	  179718 B/op	    2585 allocs/op
Benchmark_pushLogData_10_10_10_1M
Benchmark_pushLogData_10_10_10_1M-10        	    6441	    188881 ns/op	  179766 B/op	    2585 allocs/op
Benchmark_pushLogData_5_0_20_2M
Benchmark_pushLogData_5_0_20_2M-10          	   12639	     94466 ns/op	   90858 B/op	    1300 allocs/op
Benchmark_pushLogData_10_0_100_2M
Benchmark_pushLogData_10_0_100_2M-10        	    1142	   1058587 ns/op	  911518 B/op	   12998 allocs/op
Benchmark_pushLogData_50_0_200_2M
Benchmark_pushLogData_50_0_200_2M-10        	     128	   9315838 ns/op	 9217021 B/op	  129961 allocs/op
Benchmark_pushLogData_50_0_200_5M
Benchmark_pushLogData_50_0_200_5M-10        	     128	   9915882 ns/op	 9217032 B/op	  129961 allocs/op
Benchmark_pushLogData_10_1_1_1024
Benchmark_pushLogData_10_1_1_1024-10        	   64129	     18532 ns/op	   18213 B/op	     245 allocs/op
BenchmarkPushLogRecords
BenchmarkPushLogRecords-10    	  187494	      5394 ns/op	   31407 B/op	      13 allocs/op
PASS
```

Several changes were applied to the benchmarks in both before/after states:
- Disable encryption
- Setup a mock worker not to benchmark unrelated http stuff 
- Add 4 more benchmark tests that better represent real-world load with 2Mb (default) and 5Mb MaxContentLengh configuration:
  - Benchmark_pushLogData_5_0_20_2M
  - Benchmark_pushLogData_10_0_100_2M
  - Benchmark_pushLogData_50_0_200_2M
  - Benchmark_pushLogData_50_0_200_5M